Framemaker Spring Cleaning Book
Tried two different methods for inserting images and the one not taught in class is the sloppier looking one. Added paragraph elements, still need to figure out how to remove a marker, how to insert markers (which I am guessing is attached to TOC). Also did the HW requirements of list, procedure, table, image, xreference, inline formatting, lists (had third level of bullets, but ran out of time for figuring out how to change their style). Tough to learn, but like the others it seems to become more intuitive as you go...well, "intuitive" might be a stretch. Also, wanted new content...this is more click bait type stuff and really more of a reminder/motivator for me to get my spring cleaning on...not meant to be preachy:-).