Madcap Flare HW
I think I chose the wrong route of importing my Framemaker file.  After trying to clean it up, it is still not presentable and is in a state where it is neither Framemaker nor Madcap Flare ready (the template style is half there, and the chpt of FM aren't separated to topics yet).  Will be trying everything from scratch instead, but wanted to put out something.