MadCap Flare induced changes
I tried to convert Framemaker files to Flare files, and it did something to the files in this commit. I checked and didn't see anything off, so I'm guessing it's underlying code?  Also, I uploaded  my work last week with my last commit, but it's nowhere to be found. Going to try to remove then add again, but may need help.