Roffo Week 6 HW late -- three files: relative path, image, GHFM
It looks like we classmates had different responses about relative paths, so I'm hoping we can have a class review of this homework, comparing our answers to correct answers for more practice and a better understanding. If so, Thank you!
Hoping my relative path to the pelican image worked in the Markdown file.  :-)
There were 7 changed files that GitHub put here, but I unchecked 4, as they were not Week 6 HW.